+++ src/collection/cereal.rs
@@ -2,16 +2,16 @@
 
 use crate::{
     collection::{
-        recipe_tree::RecipeNode, Chain, ChainId, Profile, ProfileId, Recipe,
-        RecipeId,
+        recipe_tree::RecipeNode, Chain, ChainId, JsonBody, Profile, ProfileId,
+        Recipe, RecipeBody, RecipeId,
     },
     template::Template,
 };
 use serde::{
-    de::{Error, Visitor},
-    Deserialize, Deserializer,
+    de::{EnumAccess, Error, VariantAccess, Visitor},
+    Deserialize, Deserializer, Serialize, Serializer,
 };
-use std::hash::Hash;
+use std::{fmt::Display, hash::Hash, str::FromStr};
 
 /// A type that has an `id` field. This is ripe for a derive macro, maybe a fun
 /// project some day?
@@ -95,6 +95,17 @@ where
     Ok(map)
 }
 
+/// Deserialize a value using its `FromStr` implementation
+pub fn deserialize_from_str<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+where
+    D: Deserializer<'de>,
+    T: FromStr,
+    T::Err: Display,
+{
+    let s = String::deserialize(deserializer)?;
+    s.parse().map_err(D::Error::custom)
+}
+
 // Custom deserializer for `Template`. This is useful for deserializing values
 // that are not strings, but should be treated as strings such as numbers,
 // booleans, and nulls.
@@ -111,7 +122,7 @@ impl<'de> Deserialize<'de> for Template {
                 where
                     E: Error,
                 {
-                    Template::try_from(v.to_string()).map_err(E::custom)
+                    self.visit_string(v.to_string())
                 }
             };
         }
@@ -131,9 +142,111 @@ impl<'de> Deserialize<'de> for Template {
             visit_primitive!(visit_i64, i64);
             visit_primitive!(visit_f64, f64);
             visit_primitive!(visit_str, &str);
+
+            fn visit_string<E>(self, v: String) -> Result<Self::Value, E>
+            where
+                E: Error,
+            {
+                Template::try_from(v).map_err(E::custom)
+            }
         }
 
         deserializer.deserialize_any(TemplateVisitor)
+    }
+}
+
+impl RecipeBody {
+    // Constants for serialize/deserialization. Typically these are generated
+    // by macros, but we need custom implementation
+    const STRUCT_NAME: &'static str = "RecipeBody";
+    const VARIANT_JSON: &'static str = "json";
+    const ALL_VARIANTS: &'static [&'static str] = &[Self::VARIANT_JSON];
+}
+
+/// Custom serialization for RecipeBody, so the `Raw` variant serializes as a
+/// scalar without a tag
+impl Serialize for RecipeBody {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        // This involves a lot of duplication, but any abstraction will probably
+        // just make it worse
+        match self {
+            RecipeBody::Raw(template) => template.serialize(serializer),
+            RecipeBody::Json(value) => serializer.serialize_newtype_variant(
+                Self::STRUCT_NAME,
+                1,
+                Self::VARIANT_JSON,
+                value,
+            ),
+        }
+    }
+}
+
+// Custom deserialization for RecipeBody, to support raw template or structured
+// body with a tag
+impl<'de> Deserialize<'de> for RecipeBody {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct RecipeBodyVisitor;
+
+        /// For all primitives, parse it as a template and create a raw body
+        macro_rules! visit_primitive {
+            ($func:ident, $type:ty) => {
+                fn $func<E>(self, v: $type) -> Result<Self::Value, E>
+                where
+                    E: Error,
+                {
+                    let template =
+                        Template::try_from(v.to_string()).map_err(E::custom)?;
+                    Ok(RecipeBody::Raw(template))
+                }
+            };
+        }
+
+        impl<'de> Visitor<'de> for RecipeBodyVisitor {
+            type Value = RecipeBody;
+
+            fn expecting(
+                &self,
+                formatter: &mut std::fmt::Formatter,
+            ) -> std::fmt::Result {
+                // "!<type>" is a little wonky, but tags aren't a common YAML
+                // syntax so we should provide a hint to the user about what it
+                // means. Once they provide a tag they'll get a different error
+                // message if it's an unsupported tag
+                formatter.write_str("string, boolean, number, or tag !<type>")
+            }
+
+            visit_primitive!(visit_bool, bool);
+            visit_primitive!(visit_u64, u64);
+            visit_primitive!(visit_u128, u128);
+            visit_primitive!(visit_i64, i64);
+            visit_primitive!(visit_i128, i128);
+            visit_primitive!(visit_f64, f64);
+            visit_primitive!(visit_str, &str);
+
+            fn visit_enum<A>(self, data: A) -> Result<Self::Value, A::Error>
+            where
+                A: EnumAccess<'de>,
+            {
+                let (tag, value) = data.variant::<String>()?;
+                match tag.as_str() {
+                    RecipeBody::VARIANT_JSON => Ok(RecipeBody::Json(
+                        value.newtype_variant::<JsonBody>()?,
+                    )),
+                    other => Err(A::Error::unknown_variant(
+                        other,
+                        RecipeBody::ALL_VARIANTS,
+                    )),
+                }
+            }
+        }
+
+        deserializer.deserialize_any(RecipeBodyVisitor)
     }
 }
 
@@ -221,84 +334,23 @@ pub mod serde_duration {
         };
         Ok(Duration::from_secs(seconds))
     }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-        use rstest::rstest;
-        use serde::Serialize;
-        use serde_test::{
-            assert_de_tokens, assert_de_tokens_error, assert_ser_tokens, Token,
-        };
-
-        /// A wrapper that forces serde_test to use our custom
-        /// serialize/deserialize functions
-        #[derive(Debug, PartialEq, Serialize, Deserialize)]
-        #[serde(transparent)]
-        struct Wrap(#[serde(with = "super")] Duration);
-
-        #[rstest]
-        #[case::seconds_short(Duration::from_secs(3), "3s")]
-        #[case::seconds_long(Duration::from_secs(3000), "3000s")]
-        // Subsecond precision is lost
-        #[case::seconds_subsecond_lost(Duration::from_millis(400), "0s")]
-        #[case::seconds_subsecond_round_down(Duration::from_millis(1999), "1s")]
-        fn test_serialize(
-            #[case] duration: Duration,
-            #[case] expected: &'static str,
-        ) {
-            assert_ser_tokens(&Wrap(duration), &[Token::String(expected)]);
-        }
-
-        #[rstest]
-        #[case::seconds_zero("0s", Duration::from_secs(0))]
-        #[case::seconds_short("1s", Duration::from_secs(1))]
-        #[case::seconds_longer("100s", Duration::from_secs(100))]
-        #[case::minutes("3m", Duration::from_secs(180))]
-        #[case::hours("3h", Duration::from_secs(10800))]
-        #[case::days("2d", Duration::from_secs(172800))]
-        fn test_deserialize(
-            #[case] s: &'static str,
-            #[case] expected: Duration,
-        ) {
-            assert_de_tokens(&Wrap(expected), &[Token::Str(s)])
-        }
-
-        #[rstest]
-        #[case::negative(
-            "-1s",
-            "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
-        )]
-        #[case::whitespace(
-            " 1s ",
-            "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
-        )]
-        #[case::trailing_whitespace(
-            "1s ",
-            "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
-        )]
-        #[case::decimal(
-            "3.5s",
-            "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
-        )]
-        #[case::invalid_unit(
-            "3hr",
-            "Unknown duration unit `hr`; must be one of `s`, `m`, `h`, `d`"
-        )]
-        fn test_deserialize_error(
-            #[case] s: &'static str,
-            #[case] error: &str,
-        ) {
-            assert_de_tokens_error::<Wrap>(&[Token::Str(s)], error)
-        }
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::template::Template;
+    use super::*;
+    use crate::test_util::assert_err;
     use rstest::rstest;
-    use serde_test::{assert_de_tokens, Token};
+    use serde::Serialize;
+    use serde_json::json;
+    use serde_test::{
+        assert_de_tokens, assert_de_tokens_error, assert_ser_tokens, Token,
+    };
+    use serde_yaml::{
+        value::{Tag, TaggedValue},
+        Mapping,
+    };
+    use std::time::Duration;
 
     #[rstest]
     // boolean
@@ -316,5 +368,137 @@ mod tests {
     #[case::str_false(Token::Str("false"), "false")]
     fn test_deserialize_template(#[case] token: Token, #[case] expected: &str) {
         assert_de_tokens(&Template::from(expected), &[token]);
+    }
+
+    /// Test serializing and deserializing recipe bodies. Round trips should all
+    /// be no-ops. We use serde_yaml instead of serde_test because the handling
+    /// of enums is a bit different, and we specifically only care about YAML.
+    #[rstest]
+    #[case::raw(
+        RecipeBody::Raw("{{user_id}}".into()),
+        "{{user_id}}"
+    )]
+    #[case::json(
+        RecipeBody::Json(json!({"user": "{{user_id}}"}).into()),
+        serde_yaml::Value::Tagged(Box::new(TaggedValue{
+            tag: Tag::new("json"),
+            value: [
+                (serde_yaml::Value::from("user"), "{{user_id}}".into())
+            ].into_iter().collect::<Mapping>().into()
+        })),
+    )]
+    #[case::json_nested(
+        RecipeBody::Json(json!(r#"{"warning": "NOT an object"}"#).into()),
+        serde_yaml::Value::Tagged(Box::new(TaggedValue{
+            tag: Tag::new("json"),
+            value: r#"{"warning": "NOT an object"}"#.into()
+        })),
+    )]
+    fn test_serde_recipe_body(
+        #[case] body: RecipeBody,
+        #[case] yaml: impl Into<serde_yaml::Value>,
+    ) {
+        let yaml = yaml.into();
+        assert_eq!(
+            serde_yaml::to_value(&body).unwrap(),
+            yaml,
+            "Serialization mismatch"
+        );
+        assert_eq!(
+            serde_yaml::from_value::<RecipeBody>(yaml).unwrap(),
+            body,
+            "Deserialization mismatch"
+        );
+    }
+
+    /// Test various errors when deserializing a recipe body. We use serde_yaml
+    /// instead of serde_test because the handling of enums is a bit different,
+    /// and we specifically only care about YAML.
+    #[rstest]
+    #[case::array(
+        Vec::<i32>::new(),
+        "invalid type: sequence, expected string, boolean, number, or tag !<type>"
+    )]
+    #[case::map(
+        Mapping::default(),
+        "invalid type: map, expected string, boolean, number, or tag !<type>"
+    )]
+    // `Raw` variant is *not* accessible by tag
+    #[case::raw_tag(
+        serde_yaml::Value::Tagged(Box::new(TaggedValue{
+            tag: Tag::new("raw"),
+            value: "{{user_id}}".into()
+        })),
+        "unknown variant `raw`, expected `json`",
+    )]
+    fn test_deserialize_recipe_error(
+        #[case] yaml: impl Into<serde_yaml::Value>,
+        #[case] expected_error: &str,
+    ) {
+        assert_err!(
+            serde_yaml::from_value::<RecipeBody>(yaml.into()),
+            expected_error
+        );
+    }
+
+    /// A wrapper that forces serde_test to use our custom serialize/deserialize
+    /// functions
+    #[derive(Debug, PartialEq, Serialize, Deserialize)]
+    #[serde(transparent)]
+    struct WrapDuration(#[serde(with = "super::serde_duration")] Duration);
+
+    #[rstest]
+    #[case::seconds_short(Duration::from_secs(3), "3s")]
+    #[case::seconds_long(Duration::from_secs(3000), "3000s")]
+    // Subsecond precision is lost
+    #[case::seconds_subsecond_lost(Duration::from_millis(400), "0s")]
+    #[case::seconds_subsecond_round_down(Duration::from_millis(1999), "1s")]
+    fn test_serialize_duration(
+        #[case] duration: Duration,
+        #[case] expected: &'static str,
+    ) {
+        assert_ser_tokens(&WrapDuration(duration), &[Token::String(expected)]);
+    }
+
+    #[rstest]
+    #[case::seconds_zero("0s", Duration::from_secs(0))]
+    #[case::seconds_short("1s", Duration::from_secs(1))]
+    #[case::seconds_longer("100s", Duration::from_secs(100))]
+    #[case::minutes("3m", Duration::from_secs(180))]
+    #[case::hours("3h", Duration::from_secs(10800))]
+    #[case::days("2d", Duration::from_secs(172800))]
+    fn test_deserialize_duration(
+        #[case] s: &'static str,
+        #[case] expected: Duration,
+    ) {
+        assert_de_tokens(&WrapDuration(expected), &[Token::Str(s)])
+    }
+
+    #[rstest]
+    #[case::negative(
+        "-1s",
+        "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
+    )]
+    #[case::whitespace(
+        " 1s ",
+        "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
+    )]
+    #[case::trailing_whitespace(
+        "1s ",
+        "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
+    )]
+    #[case::decimal(
+        "3.5s",
+        "Invalid duration, must be `<quantity><unit>` (e.g. `12d`)"
+    )]
+    #[case::invalid_unit(
+        "3hr",
+        "Unknown duration unit `hr`; must be one of `s`, `m`, `h`, `d`"
+    )]
+    fn test_deserialize_duration_error(
+        #[case] s: &'static str,
+        #[case] error: &str,
+    ) {
+        assert_de_tokens_error::<WrapDuration>(&[Token::Str(s)], error)
     }
 }