+++ src/collection/models.rs
@@ -171,7 +171,7 @@ pub struct Recipe {
     /// wrong which is helpful.
     pub method: Method,
     pub url: Template,
-    pub body: Option<Template>,
+    pub body: Option<RecipeBody>,
     pub authentication: Option<Authentication>,
     #[serde(default)]
     pub query: IndexMap<String, Template>,
@@ -278,6 +278,146 @@ pub enum Authentication<T = Template> {
     Basic { username: T, password: Option<T> },
     /// `Authorization: Bearer {token}`
     Bearer(T),
+}
+
+/// Template for a request body. `Raw` is the "default" variant, which repesents
+/// a single string (parsed as a template). Other variants can be used for
+/// convenience, to construct complex bodies in common formats. The HTTP engine
+/// uses the variant to determine not only how to serialize the body, but also
+/// other parameters of the request (e.g. the `Content-Type` header).
+#[derive(Clone, Debug)]
+#[cfg_attr(test, derive(PartialEq))]
+pub enum RecipeBody {
+    Raw(Template),
+    Json(JsonBody),
+}
+
+impl RecipeBody {
+    /// For structured bodies, get the corresponding [ContentType]. Return
+    /// `None` for raw bodies
+    pub fn content_type(&self) -> Option<ContentType> {
+        match self {
+            RecipeBody::Raw(_) => None,
+            RecipeBody::Json(_) => Some(ContentType::Json),
+        }
+    }
+}
+
+#[cfg(test)]
+impl From<&str> for RecipeBody {
+    fn from(template: &str) -> Self {
+        Self::Raw(template.into())
+    }
+}
+
+/// A structured JSON recipe body. This corresponds directly to
+/// [serde_json::Value], but the value of the `String` variant is replaced with
+/// a generic param `S`. For recipes, `S = Template`, so that strings can be
+/// templatized. Other type params are used throughout the app to represent the
+/// result of certain transformations.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(test, derive(PartialEq))]
+#[serde(untagged)]
+pub enum JsonBody<S = Template> {
+    Null,
+    Bool(bool),
+    Number(serde_json::Number),
+    String(S),
+    Array(Vec<Self>),
+    Object(IndexMap<String, Self>),
+}
+
+impl<S> JsonBody<S> {
+    /// Map from `JsonBody<S>` to `JsonBody<T>` by recursively applying a
+    /// mapping function to each string value
+    pub fn map<T>(self, mapper: impl Copy + Fn(S) -> T) -> JsonBody<T> {
+        match self {
+            Self::Null => JsonBody::Null,
+            Self::Bool(b) => JsonBody::Bool(b),
+            Self::Number(n) => JsonBody::Number(n),
+            Self::String(template) => JsonBody::String(mapper(template)),
+            Self::Array(values) => JsonBody::Array(
+                values.into_iter().map(|value| value.map(mapper)).collect(),
+            ),
+            Self::Object(items) => JsonBody::Object(
+                items
+                    .into_iter()
+                    .map(|(key, value)| (key.clone(), value.map(mapper)))
+                    .collect(),
+            ),
+        }
+    }
+
+    /// Map from `&JsonBody<S>` to `JsonBody<T>` by recursively applying a
+    /// mapping function to each string value
+    pub fn map_ref<T>(&self, mapper: impl Copy + Fn(&S) -> T) -> JsonBody<T> {
+        match self {
+            Self::Null => JsonBody::Null,
+            Self::Bool(b) => JsonBody::Bool(*b),
+            Self::Number(n) => JsonBody::Number(n.clone()),
+            Self::String(template) => JsonBody::String(mapper(template)),
+            Self::Array(values) => JsonBody::Array(
+                values.iter().map(|value| value.map_ref(mapper)).collect(),
+            ),
+            Self::Object(items) => JsonBody::Object(
+                items
+                    .iter()
+                    .map(|(key, value)| (key.clone(), value.map_ref(mapper)))
+                    .collect(),
+            ),
+        }
+    }
+}
+
+/// If we're holding plain strings, we can easily convert to serde_json
+impl From<JsonBody<String>> for serde_json::Value {
+    fn from(value: JsonBody<String>) -> Self {
+        match value {
+            JsonBody::Null => serde_json::Value::Null,
+            JsonBody::Bool(b) => serde_json::Value::Bool(b),
+            JsonBody::Number(n) => serde_json::Value::Number(n.clone()),
+            JsonBody::String(s) => serde_json::Value::String(s),
+            JsonBody::Array(values) => serde_json::Value::Array(
+                values.into_iter().map(Self::from).collect(),
+            ),
+            JsonBody::Object(items) => serde_json::Value::Object(
+                items
+                    .into_iter()
+                    .map(|(key, value)| (key, value.into()))
+                    .collect(),
+            ),
+        }
+    }
+}
+
+/// Convert parsed JSON
+impl From<serde_json::Value> for JsonBody<String> {
+    fn from(value: serde_json::Value) -> Self {
+        match value {
+            serde_json::Value::Null => Self::Null,
+            serde_json::Value::Bool(b) => Self::Bool(b),
+            serde_json::Value::Number(n) => Self::Number(n),
+            serde_json::Value::String(s) => Self::String(s),
+            serde_json::Value::Array(values) => {
+                Self::Array(values.into_iter().map(Self::from).collect())
+            }
+            serde_json::Value::Object(items) => Self::Object(
+                items
+                    .into_iter()
+                    .map(|(key, value)| (key, value.into()))
+                    .collect(),
+            ),
+        }
+    }
+}
+
+/// Make it easier to construct JSON bodies in tests
+#[cfg(test)]
+impl From<serde_json::Value> for JsonBody<Template> {
+    fn from(value: serde_json::Value) -> Self {
+        let halfway: JsonBody<String> = value.into();
+        halfway.map(|s| Template::parse(s).unwrap())
+    }
 }
 
 /// A chain is a means to data from one response in another request. The chain